requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
certifi==2023.5.7
gunicorn==20.1.0